Update hunts.md
I noticed an oversight on the whitepaper on the Hunt page. The "e" was not included while attempting to spell "Get" hence I added it. Any player can start a hunt. However if a player starts a hunt without a pirate, he/she will not g"e"t any reward when finding treasures."